+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Allowed corresponding to Babel `targets.esmodules`, `targets.browsers`, `targets.node` options in `core-js-compat` and `core-js-builder`
 - Engines in compat data and results of targets parsing sorted alphabetically
 - Fixed `features/instance/match-all` entry compat data
+- Fixed `Array.prototype[@@unscopables]` descriptor (was writable)
 - Added Samsung Internet 11 compat data mapping
 
 ##### 3.5.0 - 2019.12.12